backend/requirements.txt
fastapi[all]==0.116.2
python-dotenv==1.1.1
requests==2.32.5
